Add GetByEmail and switch decimals to double
Add Produtor lookup by email: new repository interface method, repository implementation and a GET /email/{email} controller endpoint returning basic produtor info. Change numeric types for AreaHectares, Latitude and Longitude from decimal to double in models and DTOs and remove EF precision configuration in AppDbContext to align mappings. Rename several Table attributes from TB_* plural names to singular table names to match DB schema. Update connection string in appsettings.json.

+++ Dto/PropriedadeDtos.cs
@@ -6,11 +6,11 @@ public class CadastrarPropriedadeDto
     [Required] public int ProdutorId { get; set; }
     [Required] public int PlanoId { get; set; }
     [Required] public string Nome { get; set; } = null!;
-    [Required] public decimal AreaHectares { get; set; }
+    [Required] public double AreaHectares { get; set; }
     public string? Cidade { get; set; }
     [StringLength(2)] public string? Estado { get; set; }
-    public decimal? Latitude { get; set; }
-    public decimal? Longitude { get; set; }
+    public double? Latitude { get; set; }
+    public double? Longitude { get; set; }
     public string Status { get; set; } = "ATIVA";
 }
 
@@ -18,10 +18,10 @@ public class AtualizarPropriedadeDto
 {
     [Required] public int PlanoId { get; set; }
     [Required] public string Nome { get; set; } = null!;
-    [Required] public decimal AreaHectares { get; set; }
+    [Required] public double AreaHectares { get; set; }
     public string? Cidade { get; set; }
     [StringLength(2)] public string? Estado { get; set; }
-    public decimal? Latitude { get; set; }
-    public decimal? Longitude { get; set; }
+    public double? Latitude { get; set; }
+    public double? Longitude { get; set; }
     public string Status { get; set; } = "ATIVA";
 }
